backend/p2/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=backend/p2/requirements.txt backend/p2/requirements.in
#
annotated-doc==0.0.4
    # via typer
annotated-types==0.7.0
    # via pydantic
anthropic==0.84.0
    # via -r backend/p2/requirements.in
anyio==4.12.1
    # via
    #   anthropic
    #   httpx
    #   openai
bcrypt==5.0.0
    # via -r backend/p2/requirements.in
beautifulsoup4==4.14.3
    # via -r backend/p2/requirements.in
blinker==1.9.0
    # via flask
blis==1.3.3
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   duckduckgo-search
    #   flask
    #   nltk
    #   typer
cloudpathlib==0.23.0
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
cryptography==46.0.5
    # via google-auth
cymem==2.0.13
    # via
    #   preshed
    #   spacy
    #   thinc
distro==1.9.0
    # via
    #   anthropic
    #   openai
docstring-parser==0.17.0
    # via anthropic
duckduckgo-search==8.1.1
    # via -r backend/p2/requirements.in
et-xmlfile==2.0.0
    # via openpyxl
flask==3.1.3
    # via
    #   -r backend/p2/requirements.in
    #   flask-cors
flask-cors==6.0.2
    # via -r backend/p2/requirements.in
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.30.0
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.190.0
    # via google-generativeai
google-auth==2.49.0.dev0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-generativeai==0.8.6
    # via -r backend/p2/requirements.in
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.3.2
    # via playwright
grpcio==1.78.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
gunicorn==25.1.0
    # via -r backend/p2/requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   anthropic
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   spacy
jiter==0.13.0
    # via
    #   anthropic
    #   openai
joblib==1.5.3
    # via nltk
lxml==6.0.2
    # via
    #   -r backend/p2/requirements.in
    #   duckduckgo-search
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
murmurhash==1.0.15
    # via
    #   preshed
    #   spacy
    #   thinc
nltk==3.9.3
    # via textblob
numpy==2.4.2
    # via
    #   blis
    #   pandas
    #   spacy
    #   thinc
openai==2.24.0
    # via -r backend/p2/requirements.in
openpyxl==3.1.5
    # via -r backend/p2/requirements.in
packaging==26.0
    # via
    #   gunicorn
    #   pytest
    #   spacy
    #   thinc
    #   weasel
pandas==3.0.1
    # via -r backend/p2/requirements.in
playwright==1.58.0
    # via -r backend/p2/requirements.in
pluggy==1.6.0
    # via pytest
preshed==3.0.12
    # via
    #   spacy
    #   thinc
primp==1.1.1
    # via duckduckgo-search
proto-plus==1.27.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.6
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.11
    # via -r backend/p2/requirements.in
pyasn1==0.6.2
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   anthropic
    #   confection
    #   google-generativeai
    #   openai
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.41.5
    # via pydantic
pyee==13.0.1
    # via playwright
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.11.0
    # via -r backend/p2/requirements.in
pyparsing==3.3.2
    # via httplib2
pytest==9.0.2
    # via -r backend/p2/requirements.in
python-dateutil==2.9.0.post0
    # via pandas
regex==2026.2.28
    # via nltk
requests==2.32.5
    # via
    #   -r backend/p2/requirements.in
    #   google-api-core
    #   spacy
    #   weasel
rich==14.3.3
    # via typer
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
smart-open==7.5.1
    # via weasel
sniffio==1.3.1
    # via
    #   anthropic
    #   openai
soupsieve==2.8.3
    # via beautifulsoup4
spacy==3.8.11
    # via -r backend/p2/requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.5.2
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
textblob==0.19.0
    # via -r backend/p2/requirements.in
thinc==8.3.10
    # via spacy
tqdm==4.67.3
    # via
    #   google-generativeai
    #   nltk
    #   openai
    #   spacy
typer==0.24.1
    # via typer-slim
typer-slim==0.24.0
    # via
    #   spacy
    #   weasel
typing-extensions==4.15.0
    # via
    #   anthropic
    #   anyio
    #   beautifulsoup4
    #   google-generativeai
    #   grpcio
    #   openai
    #   pydantic
    #   pydantic-core
    #   pyee
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.3
    # via spacy
werkzeug==3.1.6
    # via
    #   flask
    #   flask-cors
wrapt==2.1.1
    # via smart-open

# The following packages are considered to be unsafe in a requirements file:
# setuptools
